Move DefaultFilterable into its own file
