+++ testing/tests/ui/default.stderr
@@ -81,7 +81,7 @@ error[E0277]: `Mutex<u32>` is not `|assigned_or` filterable
            and $N others
    = note: required for `&Mutex<u32>` to implement `DefaultFilterable`
 note: required by a bound in `assigned_or`
-  --> $WORKSPACE/askama/src/filters/core.rs
+  --> $WORKSPACE/askama/src/filters/default.rs
    |
    | pub fn assigned_or<L: DefaultFilterable, R>(
    |                       ^^^^^^^^^^^^^^^^^ required by this bound in `assigned_or`
